FEATURE(haproxy): Ensure service is started
 ##### SUMMARY
Check HAProxy service status at the end of the role to ensure it
is actually started. Even if the Ansible service module is called
with a "started" state, service could be down (due to binding
problem for example) and role ends without errors. The role ensure
now you end up with a started service.

 ##### IMPACT
N/A